requirements.txt
asgiref==3.4.1
Django==3.2.7
django-model-utils==4.1.1
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.2
